+++ package.json
@@ -7,8 +7,8 @@
   "bin": "dist/cli.js",
   "scripts": {
     "build": "npm run -s build:babel && npm run -s build:self",
-    "build:babel": "babel-node src/cli.js --target=node --format cjs src/{cli,index}.js",
-    "build:self": "node dist/cli.js --target=node --format cjs src/{cli,index}.js",
+    "build:babel": "babel-node src/cli.js --target=node --format cjs src/{cli,index,babel-custom,microbundle-jest}.js",
+    "build:self": "node dist/cli.js --target=node --format cjs src/{cli,index,babel-custom,microbundle-jest}.js",
     "prepare": "npm run -s build",
     "prepare:babel": "babel src/*.js -d dist && npm t",
     "lint": "eslint src",
@@ -66,6 +66,7 @@
     "@rollup/plugin-node-resolve": "^6.1.0",
     "asyncro": "^3.0.0",
     "autoprefixer": "^9.7.3",
+    "babel-jest": "^24.8.0",
     "babel-plugin-macros": "^2.8.0",
     "babel-plugin-transform-async-to-promises": "^0.8.15",
     "babel-plugin-transform-replace-expressions": "^0.2.0",
@@ -96,7 +97,6 @@
     "@babel/node": "^7.7.7",
     "@babel/plugin-proposal-optional-chaining": "^7.7.5",
     "@babel/plugin-proposal-throw-expressions": "^7.7.4",
-    "babel-jest": "^24.8.0",
     "cross-env": "^6.0.3",
     "directory-tree": "^2.2.3",
     "eslint": "^6.8.0",
