+++ src/microbundle-jest.js
@@ -1,0 +1,10 @@
+import babelJest from 'babel-jest';
+import { config } from './babel-custom';
+
+const createTransformer = (options = {}) => {
+	const babelConfig = config({}, { customOptions: { ...options, jest: true } });
+
+	return babelJest.createTransformer(babelConfig);
+};
+
+module.exports = { createTransformer };
